+++ NEWS.md
@@ -3,3 +3,4 @@
 * Improved parsing algorithm.
 * No longer allow "sub-solutions" inside stricter solutions.
 * Parser can now correctly parse sentences like "meet me on the 10th"
+* Parser can now correctly parse XX/XX dates relative to a released date.